backend/requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-environ==0.12.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
